refactor: migrate to ducks pattern(Redux Toolkit: slice.js)
