+++ src/components/CategoriesContainer.jsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from '../actions';
+} from '../slice';
 
 import { get } from '../utils';
