+++ src/components/RegionsContainer.jsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from '../actions';
+} from '../slice';
 
 import { get } from '../utils';
