+++ src/pages/RestaurantsPage.jsx
@@ -10,7 +10,7 @@ import RestaurantsContainer from '../components/RestaurantsContainer';
 
 import {
   loadInitialData,
-} from '../actions';
+} from '../slice';
 
 export default function RestaurantsPage() {
   const history = useHistory();
