+++ src/reducer.js
@@ -1,132 +1,3 @@
-import { equal } from './utils';
+import reducer from './slice';
 
-const initialReviewFields = {
-  score: '',
-  description: '',
-};
-
-const initialState = {
-  regions: [],
-  categories: [],
-  restaurants: [],
-  restaurant: null,
-  selectedRegion: null,
-  selectedCategory: null,
-  loginFields: {
-    email: '',
-    password: '',
-  },
-  accessToken: '',
-  reviewFields: {
-    ...initialReviewFields,
-  },
-};
-
-const reducers = {
-  setRegions(state, { payload: { regions } }) {
-    return {
-      ...state,
-      regions,
-    };
-  },
-
-  setCategories(state, { payload: { categories } }) {
-    return {
-      ...state,
-      categories,
-    };
-  },
-
-  setRestaurants(state, { payload: { restaurants } }) {
-    return {
-      ...state,
-      restaurants,
-    };
-  },
-
-  setRestaurant(state, { payload: { restaurant } }) {
-    return {
-      ...state,
-      restaurant,
-    };
-  },
-
-  selectRegion(state, { payload: { regionId } }) {
-    const { regions } = state;
-    return {
-      ...state,
-      selectedRegion: regions.find(equal('id', regionId)),
-    };
-  },
-
-  selectCategory(state, { payload: { categoryId } }) {
-    const { categories } = state;
-    return {
-      ...state,
-      selectedCategory: categories.find(equal('id', categoryId)),
-    };
-  },
-
-  changeLoginField(state, { payload: { name, value } }) {
-    return {
-      ...state,
-      loginFields: {
-        ...state.loginFields,
-        [name]: value,
-      },
-    };
-  },
-
-  setAccessToken(state, { payload: { accessToken } }) {
-    return {
-      ...state,
-      accessToken,
-    };
-  },
-
-  logout(state) {
-    return {
-      ...state,
-      accessToken: '',
-    };
-  },
-
-  changeReviewField(state, { payload: { name, value } }) {
-    return {
-      ...state,
-      reviewFields: {
-        ...state.reviewFields,
-        [name]: value,
-      },
-    };
-  },
-
-  clearReviewFields(state) {
-    return {
-      ...state,
-      reviewFields: {
-        ...initialReviewFields,
-      },
-    };
-  },
-
-  setReviews(state, { payload: { reviews } }) {
-    const { restaurant } = state;
-
-    return {
-      ...state,
-      restaurant: {
-        ...restaurant,
-        reviews,
-      },
-    };
-  },
-};
-
-function defaultReducer(state) {
-  return state;
-}
-
-export default function reducer(state = initialState, action) {
-  return (reducers[action.type] || defaultReducer)(state, action);
-}
+export default reducer;
